Use SetElem() to put pawn on the Goban

+++ src/gomoku/arena/goban.go
@@ -14,66 +14,66 @@ func NewGoban() *Goban {
 	return &Goban{}
 }
 
-func (goban *Goban) GetElem(row uint16, col uint16) int8 {
+func (goban *Goban) GetElem(row int32, col int32) int8 {
 	return goban[row*19+col]
 }
 
-func (goban *Goban) GetTopElem(row uint16, col uint16) int8 {
+func (goban *Goban) GetTopElem(row int32, col int32) int8 {
 	if row > 0 {
 		return goban[(row-1)*19+col]
 	}
 	return -1
 }
 
-func (goban *Goban) GetBottomElem(row uint16, col uint16) int8 {
+func (goban *Goban) GetBottomElem(row int32, col int32) int8 {
 	if row < 17 {
 		return goban[(row+1)*19+col]
 	}
 	return -1
 }
 
-func (goban *Goban) GetLeftElem(row uint16, col uint16) int8 {
+func (goban *Goban) GetLeftElem(row int32, col int32) int8 {
 	if col > 0 {
 		return goban[row*19+(col-1)]
 	}
 	return -1
 }
 
-func (goban *Goban) GetRightElem(row uint16, col uint16) int8 {
+func (goban *Goban) GetRightElem(row int32, col int32) int8 {
 	if col < 17 {
 		return goban[row*19+(col+1)]
 	}
 	return -1
 }
 
-func (goban *Goban) GetTopLeftElem(row uint16, col uint16) int8 {
+func (goban *Goban) GetTopLeftElem(row int32, col int32) int8 {
 	if row > 0 && col > 0 {
 		return goban[(row-1)*19+(col-1)]
 	}
 	return -1
 }
 
-func (goban *Goban) GetTopRightElem(row uint16, col uint16) int8 {
+func (goban *Goban) GetTopRightElem(row int32, col int32) int8 {
 	if row > 0 && col < 17 {
 		return goban[(row-1)*19+(col+1)]
 	}
 	return -1
 }
 
-func (goban *Goban) GetBottomLeftElem(row uint16, col uint16) int8 {
+func (goban *Goban) GetBottomLeftElem(row int32, col int32) int8 {
 	if row < 17 && col > 0 {
 		return goban[(row+1)*19+(col-1)]
 	}
 	return -1
 }
 
-func (goban *Goban) GetBottomRightElem(row uint16, col uint16) int8 {
+func (goban *Goban) GetBottomRightElem(row int32, col int32) int8 {
 	if row < 17 && col < 17 {
 		return goban[(row+1)*19+(col+1)]
 	}
 	return -1
 }
 
-func (goban *Goban) SetElem(row uint16, col uint16, val int8) {
+func (goban *Goban) SetElem(row int32, col int32, val int8) {
 	goban[row*19+col] = val
 }
